+++ src/main/java/com/vocalink/crossproduct/ui/dto/position/PositionDetailsTotalsDto.java
@@ -1,11 +1,11 @@
 package com.vocalink.crossproduct.ui.dto.position;
 
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import java.math.BigInteger;
 
 @Builder
 @NoArgsConstructor
@@ -13,7 +13,7 @@ import java.math.BigInteger;
 @Getter
 @Setter
 public class PositionDetailsTotalsDto {
-  private BigInteger totalCredit;
-  private BigInteger totalDebit;
-  private BigInteger totalNetPosition;
+  private BigDecimal totalCredit;
+  private BigDecimal totalDebit;
+  private BigDecimal totalNetPosition;
 }
